foodie/requirements.txt
Django==1.11
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.8.2
drfdocs==0.0.11
gunicorn==19.6.0
Markdown==2.6.11
Pillow==5.1.0
pytz==2018.4
whitenoise-3.2.1
dj-database-url-0.5.0
psycopg==2.6.2
